Prevent missing NIDAQmx from causing an error
Fixes the case when iostreaming_device_discovery() returns with error in case nidaqmx package is present but NIDAQmx drivers are missing. The intention is to skip over such error since user may actually never intend to use NI harware in his measurements. This situation was expected to be rare but is not because default YAML environment file actually suggests that nidaqmx be installed automatically on deployment of the local environment, regardless whether user installed NIDAQmx drivers and plans to use the NI hardware.