Deepcopy input_frame into _monitor_storage in StereoStreamingDeviceBase
Make a deep copy of the current input frame on each _put_monitor_frame() in order to ensure that the last frame in _monitor_storage is not changing due to reading thread writing again to the input_frame once the lock() is released. This is in order to remove risk that last portion of the monitor frame could change before FFT is finished calculating due to Python default shallow-copy behavior. Such situation would introduce unnecessary spectral leak into the analysis due to lack of _monitor_storage signal phase continuity.